Refactor the CLI tool
