+++ amt/schema/system_card.py
@@ -12,5 +12,9 @@ class SystemCard(BaseModel):
     schema_version: str = Field(default="0.1a10")
     name: str = Field(default=None)
     ai_act_profile: AiActProfile = Field(default=None)
-    selected_instruments: list[InstrumentBase] = Field(default=[])
+    provenance: dict = Field(default={})
+    description: str = Field(default=None)
+    labels: list = Field(default=[])
+    status: str = Field(default=None)
+    instruments: list[InstrumentBase] = Field(default=[])
     assessments: list[AssessmentCard] = Field(default=[])
