+++ amt/site/templates/projects/details_base.html.j2
@@ -72,7 +72,7 @@
                         <h5 class="amt-widget-title rvo-link--with-icon"><span
                             class="utrecht-icon rvo-icon rvo-icon-permanent-beta rvo-icon--md rvo-icon--hemelblauw rvo-link__icon--before"
                             role="img" aria-label=""></span>Welke instrumenten zijn uitgevoerd?</h5>
-                        <a class="rvo-text--bold rvo-text--subtle">1/3 uitgevoerd</a>
+                        <a class="rvo-text--bold rvo-text--subtle">{{ instrument_state.count_0 }}/ {{ instrument_state.count_1 }} uitgevoerd</a>
                     </div>
                     <ul class="rvo-item-list">
                         <li class="rvo-item-list__item">
