+++ amt/site/templates/projects/details_info.html.j2
@@ -21,7 +21,9 @@
             </tr>
             <tr class="rvo-table-row">
                 <td class="rvo-table-cell">Levenscyclus</td>
-                <td class="rvo-table-cell"><a class="rvo-tag rvo-tag--active rvo-tag--default rvo-tag--success"><span
+                <td class="rvo-table-cell">
+{#                    TODO: align production and vinkje in centre #}
+                    <a class="rvo-tag rvo-tag--active rvo-tag--default rvo-tag--success"><span
     class="utrecht-icon rvo-icon rvo-icon-bevestiging rvo-icon--lg"
     role="img"
     aria-label="Bevestiging"
